Update service name to include businsess
"1. So that it’s clear on every page that this service is currently only for business studies ITT. The banner is the best place to do this (e.g. we're also trying to do it here - https://dfedigital.atlassian.net/secure/RapidBoard.jspa?rapidView=2&projectKey=BATSC&modal=detail&selectedIssue=BATSA-232 - but I'm not sure that's the right option')
2. The word ‘other’ in banner text ‘Search for other teacher training courses on UCAS’ doesn’t always make sense, unless we update the service name.
"