Display a minimum about section on courses
When a course doesn't have an about section on it, but the provider
does have a website defined in their contact details, we should
display a minimum "About the course" section highlighting the website.

As discussed with @fofr, we should hide the website from the main
contact details section in this situation to avoid duplicating it.

We should also not display it when the user is in preview mode.