Cope gracefully with api returning null for courses
The api shouldn't be returning null for the filter I'm looking at but seeing as it does let's get rid of the not very helpful null ref exception. It seems reasonable to interpret a null response for courses from the api as meaning there are zero courses.

Note that this doesn't fix the reported issue as the link says 24 courses but the api isn't returning any. Will investigate further.