Add external link analytics
This is a stripped down version of the external link tracker from
govuk_frontend_toolkit[0]. A notable difference is this module doesn't
currently work with links that are dynamically inserted into the DOM
after page load. Because our application is fully server-side rendered
at the moment, this can be remedied later if necessary.

- [0]: https://github.com/alphagov/govuk_frontend_toolkit/blob/master/javascripts/govuk/analytics/external-link-tracker.js