Fix commited merge conflicts
These were not picked up by the tests because we don't have tests for
views.

These should be picked up by the smoke tests on the dev env.

These did get picked up by the review app which was throwing a Heroku error,
but I did not realise it was because of this and did not investigate
at the time.