Reduce importance of training location codes
When applying you don't need to enter these codes, only the name of the location.

A code is shown once a location is selected on the application form but not before.

A code might be useful for confirmation once a choice is made in an application, so leave the code in but give it less prominence.