Improve how the Google geocoder is called
Before this change, the geocoder call and the maps autocomplete calls
were using different mechanisms for limiting the results to the UK.

The autocomplete set 'components=country:uk' which only returned UK results.
The call to the geocoder involved a 'region=uk' call that seems to still return
results outside the UK (e.g. if you search for New York, it returns the one in the US
rather than the one in Lincolnshire), and this meant the code had to have a hack around
this case.

This commit updates the geocoder call with the same 'components' query param, rendering
the hack unnecessary.