[1162] Log swallowed exceptions from suggested search
This is a follow-up to an edge-case failure in production that resulted
in the whole results page failing. Having patched it to swallow
exceptions to improve the user experience we now want to be made aware
if this is happening so we can take action. These logs can be pushed
into app insights etc to create alerting.