[1164] Fix 500 error on the qualifications page
The offending URL contains separate `qualification` parameters as
opposed to the comma separated `qualifications` string that the POST
uses now. The frontend template can't deal with it, the POST knows
how to deal with it, so just copy over the logic into the GET. Yes,
this is a bit of repetition but it's only the second instance and I
think it's fine.